+++ bin/cli.js
@@ -11,9 +11,33 @@ process.env.PIRAL_DOCS_BASE_DIR = baseDir;
 
 loadPlugins();
 
-apps.buildPiral(baseDir, {
-  entry: `${relative(baseDir, __dirname)}/../src/index.pug`,
-  target: `${outputPath}/index.html`,
-  type: 'release',
-  bundlerName: 'parcel',
-}).then(() => process.exit(0), () => process.exit(1));
+const entry = `${relative(baseDir, __dirname)}/../src/index.pug`;
+const target = `${outputPath}/index.html`;
+const bundlerName = 'parcel';
+
+switch (process.argv.pop()) {
+  case 'watch':
+    return apps
+      .debugPiral(baseDir, {
+        entry,
+        target,
+        bundlerName,
+      })
+      .then(
+        () => process.exit(0),
+        () => process.exit(1),
+      );
+  case 'build':
+  default:
+    return apps
+      .buildPiral(baseDir, {
+        entry,
+        target,
+        bundlerName,
+        type: 'release',
+      })
+      .then(
+        () => process.exit(0),
+        () => process.exit(1),
+      );
+}